AMD Host System Management Port driver
Updates to driver code based on internal feedback reviews.

The Host System Management Port (HSMP) driver provides userspace
and kernel access to system management functions on AMD systems.
The HSMP Port is defined in the Family 19h PPR [1].

The HSMP port is accessed through a set of SMN registers that allow
for setting and retrieving system management settings. Both setting
and retrieving require writing a set of request values and then
reading one or more responses once firmware has responded. The HSMP
provides a kernel interface to allow easy setting and reading
of the hsmp interfaces and a userspace interface via sysfs for
monitoring applications to use.

[1] AMD PPR for Family 19h, Model 01h, Revision B1 (Chapter 7)
https://www.amd.com/system/files/TechDocs/55898_B1_pub_0.50.zip

Co-developed-by: Lewis Carroll <lewis.carroll@amd.com>

---

Updates for Version 3:
- Squash calculation of tctl into a single statement
- Reduce the use of pr_debug()
- Replace EXPORT_SYMBOL with EXPORT_SYMBOL_GPL
- Re-write of the hsmp node and socket initialization to allocate
  node and socket arrays based on the number of nodes and sockets
  present.
- Updated raw interface creation to allocate array based on number
  sockets present instead of static size.
- Updates sysfs initialization to fail driver load on any sysfs
  directory or file creation.
- The previous two updates required the addition of a hsmp_cleanup()
  routine to properly cleanup the driver on a failed load or at exit.
- Corrected patch sign-offs
- Removed pr_err() output from interface calls, the failure return
  code should be enough for failure recognition.
- Added helper routines hsmp_get_one_value() and hsmp_set_one_variable()
  for the majority of hsmp interfaces that only get or set one value.
- Corrected arg to set boost limit interfaces to take a u16, the HSMP
  interface limits the value to 16 bits.

Updates for Version 2:
- Removed the use of likely() and unlikely()
- Updated the Copyright dates in amd_hsmp.c and amd_hsmp.h
- Moved the module parameter declarations to the top of the file
- Better organize global declarations in amd_hsmp.c
- Updated the PPR link in commit message to include link to actual
  PPR and include version name and Chapter for HSMP.
- Re-write loop waiting for firmware response to HSMP messages
  based on Tom's suggestions.
- Update nbio bus limit initialization to simplify the loop.
- Use object copy when sorting nbios instead of copying individual
  object members.
- Break up hsmp_send_message() to provide a wrapper that does the
  preliminary checks and takes the lock to call __hsmp_send_msg().
  This also includes adding a routine to parse errors from sending
  hsmp messages.
- Add 'depends on CPU_SUP_AMD' to Kconfig for amd hsmp driver.

+++ amd_hsmp.h
@@ -1,8 +1,8 @@
 /* SPDX-License-Identifier: GPL-2.0 */
 /*
- * Copyright (C) 2007-2019 Advanced Micro Devices, Inc.
+ * Copyright (C) 2021 Advanced Micro Devices, Inc.
+ * Author: Nathan Fontenot <nathan.fontenot@amd.com>
  * Author: Lewis Carroll <lewis.carroll@amd.com>
- * Maintainer: Nathan Fontenot <nathan.fontenot@amd.com>
  */
 
 /*
@@ -90,20 +90,23 @@ int hsmp_get_power_limit_max(int socket_id, u32 *limit_mw);
 
 /*
  * Set HSMP boost limit for the system.
+ * The boost limit is restricted to a 16-bit value by the HSMP interface.
  */
-int hsmp_set_boost_limit_system(u32 limit_mhz);
+int hsmp_set_boost_limit_system(u16 limit_mhz);
 
 /*
  * Set HSMP boost limit for a specific core.
+ * The boost limit is restricted to a 16-bit value by the HSMP interface.
  * Returns -ENODEV if the specified CPU does not exist.
  */
-int hsmp_set_boost_limit_cpu(int cpu, u32 limit_mhz);
+int hsmp_set_boost_limit_cpu(int cpu, u16 limit_mhz);
 
 /*
  * Set HSMP boost limit for all cores in the specified socket.
+ * The boost limit is restricted to a 16-bit value by the HSMP interface.
  * Returns -ENODEV if the specified socket does not exist.
  */
-int hsmp_set_boost_limit_socket(int socket_id, u32 limit_mhz);
+int hsmp_set_boost_limit_socket(int socket_id, u16 limit_mhz);
 
 /*
  * Get HSMP boost limit for a specific core.
